+++ app/typescript/controllers/slider_controller.ts
@@ -1,7 +1,7 @@
 import { Controller } from '@hotwired/stimulus'
 import noUiSlider, { API, Options } from 'nouislider'
 
-const SliderController = class extends Controller<HTMLElement> {
+const Slider = class extends Controller<HTMLElement> {
   static name = 'slider'
 
   // targets
@@ -110,7 +110,4 @@ const SliderController = class extends Controller<HTMLElement> {
   }
 }
 
-type Slider = InstanceType<typeof SliderController>
-
-export { SliderController }
-export type { Slider }
+export default Slider