+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 ## [Unreleased] - YYYY-MM-DD
+
 ### Added
+
+- Helper function to load stimulus JS controllers
+- Helper function to lazy load stimulus JS controllers with [stimulus-controller-resolver](https://github.com/danieldiekmeier/stimulus-controller-resolver)
+
 ### Changed
+
+- Convert stimulus JS controllers to default export
+
 ### Deprecated
+
 ### Removed
+
 ### Fixed
-- Accordion: hide content on initialize to prevent FOUC
+
+- Hide accordion content on initialize to prevent [FOUC](https://en.wikipedia.org/wiki/Flash_of_unstyled_content)
+
 ### Security
 
 ## [1.0.0] - 2025-08-04